Extract configuration to individual object
config: extract the configuration to individual object with methods (*);
config: switch function documentation to sphinx-friendly format;
config: fix two codacy issues;
commands: refactor the usage of config;
pcs: refactor the usage of config.

(*) This way the working with config is much easier and moreover, this step
serves as a precursor to temporary config.